Repair readability after main refresh
