+++ crates/temper-observe/src/otel.rs
@@ -19,15 +19,11 @@
 //! | `TEMPER_TRACE_QUEUE_SIZE` | Max buffered spans before drop (default: 2048, range: 128–32768) |
 //! | `TEMPER_LOG_QUEUE_SIZE` | Max buffered log records before drop (default: 2048, range: 128–32768) |
 
-use std::collections::HashMap;
 use std::sync::OnceLock;
 use std::time::Duration;
 
 use opentelemetry::KeyValue;
-use opentelemetry::metrics::{Counter, Gauge};
-use opentelemetry::trace::{
-    Link, SamplingDecision, SamplingResult, SpanKind, TraceId, TraceState, TracerProvider as _,
-};
+use opentelemetry::trace::TracerProvider as _;
 use opentelemetry_appender_tracing::layer::OpenTelemetryTracingBridge;
 use opentelemetry_otlp::{LogExporter, MetricExporter, SpanExporter, WithExportConfig};
 use opentelemetry_sdk::Resource;
@@ -37,265 +33,22 @@ use opentelemetry_sdk::logs::{
 use opentelemetry_sdk::metrics::{PeriodicReader, SdkMeterProvider};
 use opentelemetry_sdk::trace::{
     BatchConfigBuilder as SpanBatchConfigBuilder, BatchSpanProcessor, Sampler, SdkTracerProvider,
-    ShouldSample,
 };
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::prelude::*;
 
-/// Span names dropped at the sampler layer — trace volume noise with no
-/// operator value. See ADR-0052/0053 (instrumentation policy + hygiene).
-///
-/// - `turso.configured_connection`: DB pool internal; no timing value.
-/// - `clock_time_get`: WASI guest syscall; fires per-LLM-token in high volume.
-const DROPPED_SPAN_NAMES: &[&str] = &["turso.configured_connection", "clock_time_get"];
+mod config;
+mod sampler;
 
-const WASM_AUXILIARY_SAMPLE_RATE_ENV: &str = "TEMPER_TRACE_WASM_AUX_SAMPLE_PCT";
-const DISPATCH_BACKGROUND_SAMPLE_RATE_ENV: &str = "TEMPER_TRACE_DISPATCH_BACKGROUND_SAMPLE_PCT";
-const WASM_AUXILIARY_SAMPLE_RATE_DEFAULT: u8 = 5;
-const DISPATCH_BACKGROUND_SAMPLE_RATE_DEFAULT: u8 = 25;
-/// Span name prefixes that are sampled at a reduced rate. Do not include
-/// tool/provider guest module boundaries here: guest spans inherit parent
-/// sampling, so reducing those boundaries can sever traces operators need when
-/// following work through WASM.
-const WASM_AUXILIARY_PREFIXES: &[&str] = &["wasm:workspace_fs", "wasm:monty_repl"];
-const DISPATCH_BACKGROUND_PREFIXES: &[&str] = &[
-    "dispatch.phase.query_projection",
-    "dispatch.background_wasm_integrations",
-    "dispatch.scheduled_actions",
-];
+use config::{
+    parse_otlp_headers, read_non_empty_env, resolve_deployment_environment, resolve_otel_config,
+    resolve_service_version,
+};
+use sampler::{
+    DISPATCH_BACKGROUND_SAMPLE_RATE_DEFAULT, NameBasedSampler, TraceSamplerConfig,
+    WASM_AUXILIARY_SAMPLE_RATE_DEFAULT, record_trace_sampler_config,
+};
 
-#[derive(Clone, Debug)]
-struct ReducedSampleRule {
-    name: &'static str,
-    prefixes: &'static [&'static str],
-    sample_rate_pct: u8,
-}
-
-impl ReducedSampleRule {
-    fn matches(&self, span_name: &str) -> bool {
-        self.prefixes
-            .iter()
-            .any(|prefix| span_name.starts_with(prefix))
-    }
-}
-
-#[derive(Clone, Debug)]
-struct TraceSamplerConfig {
-    reduced_rules: Vec<ReducedSampleRule>,
-}
-
-impl Default for TraceSamplerConfig {
-    fn default() -> Self {
-        Self::with_rates(
-            WASM_AUXILIARY_SAMPLE_RATE_DEFAULT,
-            DISPATCH_BACKGROUND_SAMPLE_RATE_DEFAULT,
-        )
-    }
-}
-
-impl TraceSamplerConfig {
-    fn from_env() -> Self {
-        Self::with_rates(
-            sample_rate_from_env(
-                WASM_AUXILIARY_SAMPLE_RATE_ENV,
-                WASM_AUXILIARY_SAMPLE_RATE_DEFAULT,
-            ),
-            sample_rate_from_env(
-                DISPATCH_BACKGROUND_SAMPLE_RATE_ENV,
-                DISPATCH_BACKGROUND_SAMPLE_RATE_DEFAULT,
-            ),
-        )
-    }
-
-    fn with_rates(
-        wasm_auxiliary_sample_rate_pct: u8,
-        dispatch_background_sample_rate_pct: u8,
-    ) -> Self {
-        Self {
-            reduced_rules: vec![
-                ReducedSampleRule {
-                    name: "wasm_auxiliary",
-                    prefixes: WASM_AUXILIARY_PREFIXES,
-                    sample_rate_pct: wasm_auxiliary_sample_rate_pct,
-                },
-                ReducedSampleRule {
-                    name: "dispatch_background",
-                    prefixes: DISPATCH_BACKGROUND_PREFIXES,
-                    sample_rate_pct: dispatch_background_sample_rate_pct,
-                },
-            ],
-        }
-    }
-
-    fn reduced_rule_rate(&self, rule_name: &str) -> Option<u8> {
-        self.reduced_rules
-            .iter()
-            .find(|rule| rule.name == rule_name)
-            .map(|rule| rule.sample_rate_pct)
-    }
-
-    fn reduced_prefix_count(&self) -> usize {
-        self.reduced_rules
-            .iter()
-            .map(|rule| rule.prefixes.len())
-            .sum()
-    }
-}
-
-struct TraceSamplerMetrics {
-    decisions_total: Counter<u64>,
-    configured_rules: Gauge<u64>,
-    reduced_sample_rate_pct: Gauge<u64>,
-}
-
-fn trace_sampler_metrics() -> &'static TraceSamplerMetrics {
-    static METRICS: OnceLock<TraceSamplerMetrics> = OnceLock::new();
-    METRICS.get_or_init(|| {
-        let meter = opentelemetry::global::meter("temper.observe");
-        TraceSamplerMetrics {
-            decisions_total: meter
-                .u64_counter("temper_trace_sampler_decisions_total")
-                .with_description(
-                    "Trace sampler decisions by bounded rule name and final sampling decision.",
-                )
-                .build(),
-            configured_rules: meter
-                .u64_gauge("temper_trace_sampler_configured_rules")
-                .with_description(
-                    "Configured trace sampler rules by rule kind. Values are counts, not rates.",
-                )
-                .build(),
-            reduced_sample_rate_pct: meter
-                .u64_gauge("temper_trace_sampler_reduced_sample_rate_pct")
-                .with_unit("%")
-                .with_description("Effective trace sampler keep rate for reduced-prefix rules.")
-                .build(),
-        }
-    })
-}
-
-fn sample_rate_from_env(var: &str, default: u8) -> u8 {
-    read_non_empty_env(var)
-        .and_then(|value| value.parse::<u16>().ok())
-        .map(|value| value.min(100) as u8)
-        .unwrap_or(default)
-}
-
-fn sampling_decision_label(decision: &SamplingDecision) -> &'static str {
-    match decision {
-        SamplingDecision::Drop => "drop",
-        SamplingDecision::RecordOnly => "record_only",
-        SamplingDecision::RecordAndSample => "record_and_sample",
-    }
-}
-
-fn record_trace_sampler_decision(rule: &'static str, decision: &'static str) {
-    trace_sampler_metrics().decisions_total.add(
-        1,
-        &[
-            KeyValue::new("rule", rule),
-            KeyValue::new("decision", decision),
-        ],
-    );
-}
-
-fn record_trace_sampler_config(config: &TraceSamplerConfig) {
-    trace_sampler_metrics().configured_rules.record(
-        DROPPED_SPAN_NAMES.len() as u64,
-        &[KeyValue::new("rule_kind", "drop_exact")],
-    );
-    trace_sampler_metrics().configured_rules.record(
-        config.reduced_prefix_count() as u64,
-        &[KeyValue::new("rule_kind", "reduced_prefix")],
-    );
-    for rule in &config.reduced_rules {
-        trace_sampler_metrics().reduced_sample_rate_pct.record(
-            rule.sample_rate_pct as u64,
-            &[KeyValue::new("rule", rule.name)],
-        );
-    }
-}
-
-/// Trace-id-based deterministic sampling: keep if `(trace_id_low_u64 % 100) < rate_pct`.
-/// Using trace_id instead of random keeps the decision consistent across
-/// span siblings in the same trace, so a sampled trace keeps all its spans.
-fn trace_id_sample_at(trace_id: TraceId, rate_pct: u8) -> bool {
-    let bytes = trace_id.to_bytes();
-    // Low 8 bytes as u64.
-    let mut low = 0u64;
-    for b in &bytes[8..16] {
-        low = (low << 8) | (*b as u64);
-    }
-    (low % 100) < rate_pct as u64
-}
-
-/// Sampler that drops specific span names outright, reduced-samples specific
-/// prefixes, and delegates every other span decision to the wrapped inner
-/// sampler (default: parent-based AlwaysOn).
-///
-/// Implemented here rather than configured via `OTEL_TRACES_SAMPLER_ARG` so
-/// the drop rules live in source (grep-able) and survive env-var rewrites by
-/// tenant/deploy tooling.
-#[derive(Debug, Clone)]
-struct NameBasedSampler {
-    inner: Sampler,
-    config: TraceSamplerConfig,
-}
-
-impl ShouldSample for NameBasedSampler {
-    fn should_sample(
-        &self,
-        parent_context: Option<&opentelemetry::Context>,
-        trace_id: TraceId,
-        name: &str,
-        span_kind: &SpanKind,
-        attributes: &[KeyValue],
-        links: &[Link],
-    ) -> SamplingResult {
-        if DROPPED_SPAN_NAMES.contains(&name) {
-            record_trace_sampler_decision("drop_exact", "drop");
-            return SamplingResult {
-                decision: SamplingDecision::Drop,
-                attributes: Vec::new(),
-                trace_state: TraceState::default(),
-            };
-        }
-        if let Some(rule) = self
-            .config
-            .reduced_rules
-            .iter()
-            .find(|rule| rule.matches(name))
-        {
-            if !trace_id_sample_at(trace_id, rule.sample_rate_pct) {
-                record_trace_sampler_decision(rule.name, "reduced_drop");
-                return SamplingResult {
-                    decision: SamplingDecision::Drop,
-                    attributes: Vec::new(),
-                    trace_state: TraceState::default(),
-                };
-            }
-            let result = self.inner.should_sample(
-                parent_context,
-                trace_id,
-                name,
-                span_kind,
-                attributes,
-                links,
-            );
-            record_trace_sampler_decision(rule.name, sampling_decision_label(&result.decision));
-            return result;
-        }
-
-        let result =
-            self.inner
-                .should_sample(parent_context, trace_id, name, span_kind, attributes, links);
-        record_trace_sampler_decision("delegate", sampling_decision_label(&result.decision));
-        result
-    }
-}
-
-/// Default OTLP endpoint for Logfire.
-const LOGFIRE_ENDPOINT: &str = "https://logfire-us.pydantic.dev";
 const OTEL_EXPORTER_BUILD_RETRY_ATTEMPTS: usize = 3;
 const OTEL_EXPORTER_RETRY_BASE_DELAY_MS: u64 = 250;
 const TRACE_BATCH_MAX_QUEUE_SIZE: usize = 2_048;
@@ -316,37 +69,6 @@ fn queue_size_from_env(var: &str, default: usize) -> usize {
         .clamp(128, 32_768)
 }
 
-#[derive(Clone, Copy, Debug)]
-enum EndpointSource {
-    OtlpEndpoint,
-    OtlpExporterEndpoint,
-    LogfireToken,
-}
-
-impl EndpointSource {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::OtlpEndpoint => "OTLP_ENDPOINT",
-            Self::OtlpExporterEndpoint => "OTEL_EXPORTER_OTLP_ENDPOINT",
-            Self::LogfireToken => "LOGFIRE_TOKEN",
-        }
-    }
-}
-
-#[derive(Debug)]
-struct ResolvedOtelConfig {
-    endpoint: String,
-    endpoint_source: EndpointSource,
-    logfire_token: Option<String>,
-}
-
-fn read_non_empty_env(var_name: &str) -> Option<String> {
-    std::env::var(var_name)
-        .ok()
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
 /// Process-lifetime runtime id shared by traces and profiler uploads.
 ///
 /// Datadog uses this resource/tag value to stitch profiles back to the APM
@@ -358,83 +80,6 @@ pub fn runtime_id() -> &'static str {
             read_non_empty_env("DD_RUNTIME_ID").unwrap_or_else(|| uuid::Uuid::new_v4().to_string())
         })
         .as_str()
-}
-
-fn parse_otlp_headers(raw: &str) -> HashMap<String, String> {
-    raw.split(',')
-        .filter_map(|pair| pair.split_once('='))
-        .map(|(key, value)| (key.trim().to_string(), value.trim().to_string()))
-        .filter(|(key, value)| !key.is_empty() && !value.is_empty())
-        .collect()
-}
-
-fn resolve_deployment_environment() -> Option<String> {
-    if let Some(environment) = read_non_empty_env("DD_ENV") {
-        return Some(environment);
-    }
-
-    if let Some(environment) = read_non_empty_env("LOGFIRE_ENVIRONMENT") {
-        return Some(environment);
-    }
-
-    let resource_attrs = read_non_empty_env("OTEL_RESOURCE_ATTRIBUTES")?;
-    for raw_pair in resource_attrs.split(',') {
-        let Some((key, value)) = raw_pair.split_once('=') else {
-            continue;
-        };
-        if key.trim() == "deployment.environment.name" {
-            let trimmed = value.trim();
-            if !trimmed.is_empty() {
-                return Some(trimmed.to_string());
-            }
-        }
-    }
-    None
-}
-
-fn resolve_service_version() -> Option<String> {
-    read_non_empty_env("DD_VERSION")
-}
-
-fn resolve_otel_config() -> Option<ResolvedOtelConfig> {
-    let otlp_endpoint = read_non_empty_env("OTLP_ENDPOINT");
-    let otel_exporter_endpoint = read_non_empty_env("OTEL_EXPORTER_OTLP_ENDPOINT");
-    let logfire_token = read_non_empty_env("LOGFIRE_TOKEN");
-
-    if std::env::var_os("OTLP_ENDPOINT").is_some() && otlp_endpoint.is_none() {
-        eprintln!("OTLP_ENDPOINT is set but empty; ignoring it.");
-    }
-    if std::env::var_os("OTEL_EXPORTER_OTLP_ENDPOINT").is_some() && otel_exporter_endpoint.is_none()
-    {
-        eprintln!("OTEL_EXPORTER_OTLP_ENDPOINT is set but empty; ignoring it.");
-    }
-    if std::env::var_os("LOGFIRE_TOKEN").is_some() && logfire_token.is_none() {
-        eprintln!("LOGFIRE_TOKEN is set but empty; skipping Authorization header.");
-    }
-
-    if let (Some(otlp), Some(otel_exporter)) = (&otlp_endpoint, &otel_exporter_endpoint)
-        && otlp != otel_exporter
-    {
-        eprintln!(
-            "Both OTLP_ENDPOINT and OTEL_EXPORTER_OTLP_ENDPOINT are set. Using OTLP_ENDPOINT."
-        );
-    }
-
-    let (endpoint, endpoint_source) = if let Some(endpoint) = otlp_endpoint {
-        (endpoint, EndpointSource::OtlpEndpoint)
-    } else if let Some(endpoint) = otel_exporter_endpoint {
-        (endpoint, EndpointSource::OtlpExporterEndpoint)
-    } else if logfire_token.is_some() {
-        (LOGFIRE_ENDPOINT.to_string(), EndpointSource::LogfireToken)
-    } else {
-        return None;
-    };
-
-    Some(ResolvedOtelConfig {
-        endpoint,
-        endpoint_source,
-        logfire_token,
-    })
 }
 
 fn build_with_retry<T, E, F>(component: &str, mut build: F) -> Result<T, Box<dyn std::error::Error>>
@@ -803,223 +448,4 @@ pub fn init_stderr_only() {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::Mutex;
-
-    static ENV_LOCK: Mutex<()> = Mutex::new(());
-    const TEST_ENV_VARS: [&str; 5] = [
-        "OTLP_ENDPOINT",
-        "OTEL_EXPORTER_OTLP_ENDPOINT",
-        "LOGFIRE_TOKEN",
-        "TEMPER_TRACE_WASM_AUX_SAMPLE_PCT",
-        "TEMPER_TRACE_DISPATCH_BACKGROUND_SAMPLE_PCT",
-    ];
-
-    fn with_test_env(values: &[(&str, Option<&str>)], f: impl FnOnce()) {
-        let _guard = ENV_LOCK.lock().expect("env mutex must lock");
-        let snapshot: Vec<(&str, Option<String>)> = TEST_ENV_VARS
-            .iter()
-            .map(|key| (*key, std::env::var(key).ok()))
-            .collect();
-
-        for (key, value) in values {
-            match value {
-                Some(v) => unsafe { std::env::set_var(key, v) },
-                None => unsafe { std::env::remove_var(key) },
-            }
-        }
-
-        f();
-
-        for (key, value) in snapshot {
-            match value {
-                Some(v) => unsafe { std::env::set_var(key, v) },
-                None => unsafe { std::env::remove_var(key) },
-            }
-        }
-    }
-
-    #[test]
-    fn resolve_config_prefers_otlp_endpoint() {
-        with_test_env(
-            &[
-                ("OTLP_ENDPOINT", Some("http://otlp:4318")),
-                ("OTEL_EXPORTER_OTLP_ENDPOINT", Some("http://other:4318")),
-                ("LOGFIRE_TOKEN", Some("abc123")),
-            ],
-            || {
-                let config = resolve_otel_config().expect("config should resolve");
-                assert_eq!(config.endpoint, "http://otlp:4318");
-                assert_eq!(config.endpoint_source.as_str(), "OTLP_ENDPOINT");
-                assert_eq!(config.logfire_token.as_deref(), Some("abc123"));
-            },
-        );
-    }
-
-    #[test]
-    fn resolve_config_uses_exporter_endpoint_when_otlp_missing() {
-        with_test_env(
-            &[
-                ("OTLP_ENDPOINT", None),
-                ("OTEL_EXPORTER_OTLP_ENDPOINT", Some("http://collector:4318")),
-                ("LOGFIRE_TOKEN", None),
-            ],
-            || {
-                let config = resolve_otel_config().expect("config should resolve");
-                assert_eq!(config.endpoint, "http://collector:4318");
-                assert_eq!(
-                    config.endpoint_source.as_str(),
-                    "OTEL_EXPORTER_OTLP_ENDPOINT"
-                );
-                assert_eq!(config.logfire_token, None);
-            },
-        );
-    }
-
-    #[test]
-    fn resolve_config_falls_back_to_logfire_token() {
-        with_test_env(
-            &[
-                ("OTLP_ENDPOINT", None),
-                ("OTEL_EXPORTER_OTLP_ENDPOINT", None),
-                ("LOGFIRE_TOKEN", Some("abc123")),
-            ],
-            || {
-                let config = resolve_otel_config().expect("config should resolve");
-                assert_eq!(config.endpoint, LOGFIRE_ENDPOINT);
-                assert_eq!(config.endpoint_source.as_str(), "LOGFIRE_TOKEN");
-                assert_eq!(config.logfire_token.as_deref(), Some("abc123"));
-            },
-        );
-    }
-
-    #[test]
-    fn name_based_sampler_drops_configured_span_names() {
-        use opentelemetry::trace::TraceId;
-        let sampler = NameBasedSampler {
-            inner: Sampler::AlwaysOn,
-            config: TraceSamplerConfig::default(),
-        };
-        let trace_id = TraceId::from_bytes([0u8; 16]);
-
-        for dropped in ["turso.configured_connection", "clock_time_get"] {
-            let result =
-                sampler.should_sample(None, trace_id, dropped, &SpanKind::Internal, &[], &[]);
-            assert!(
-                matches!(result.decision, SamplingDecision::Drop),
-                "sampler must drop {dropped}",
-            );
-        }
-    }
-
-    #[test]
-    fn reduced_sample_prefixes_keep_roughly_5_percent() {
-        let sampler = NameBasedSampler {
-            inner: Sampler::AlwaysOn,
-            config: TraceSamplerConfig::default(),
-        };
-        let mut kept = 0usize;
-        let total = 10_000usize;
-        for i in 0..total {
-            let mut bytes = [0u8; 16];
-            bytes[8..].copy_from_slice(&(i as u64).to_be_bytes());
-            let trace_id = TraceId::from_bytes(bytes);
-            let r = sampler.should_sample(
-                None,
-                trace_id,
-                "wasm:workspace_fs.read",
-                &SpanKind::Internal,
-                &[],
-                &[],
-            );
-            if matches!(r.decision, SamplingDecision::RecordAndSample) {
-                kept += 1;
-            }
-        }
-        let kept_pct = (kept as f64 / total as f64) * 100.0;
-        assert!(
-            (3.0..=7.0).contains(&kept_pct),
-            "reduced sample should keep ~5%, got {kept_pct:.1}%"
-        );
-    }
-
-    #[test]
-    fn monty_repl_boundary_is_not_reduced_sampled() {
-        use opentelemetry::trace::TraceId;
-        let sampler = NameBasedSampler {
-            inner: Sampler::AlwaysOn,
-            config: TraceSamplerConfig::default(),
-        };
-        let trace_id = TraceId::from_bytes([0u8; 16]);
-        let result = sampler.should_sample(
-            None,
-            trace_id,
-            "wasm:monty_repl",
-            &SpanKind::Internal,
-            &[],
-            &[],
-        );
-
-        assert!(
-            matches!(result.decision, SamplingDecision::RecordAndSample),
-            "monty_repl must keep full sampling so guest tool spans stay stitched"
-        );
-    }
-
-    #[test]
-    fn name_based_sampler_delegates_other_spans() {
-        use opentelemetry::trace::TraceId;
-        let sampler = NameBasedSampler {
-            inner: Sampler::AlwaysOn,
-            config: TraceSamplerConfig::default(),
-        };
-        let trace_id = TraceId::from_bytes([0u8; 16]);
-        let result = sampler.should_sample(
-            None,
-            trace_id,
-            "dispatch.dispatch_tenant_action_core",
-            &SpanKind::Internal,
-            &[],
-            &[],
-        );
-        assert!(
-            matches!(result.decision, SamplingDecision::RecordAndSample),
-            "non-dropped span must keep AlwaysOn decision",
-        );
-    }
-
-    #[test]
-    fn resolve_config_ignores_empty_values() {
-        with_test_env(
-            &[
-                ("OTLP_ENDPOINT", Some("   ")),
-                ("OTEL_EXPORTER_OTLP_ENDPOINT", Some("")),
-                ("LOGFIRE_TOKEN", Some(" ")),
-            ],
-            || {
-                let config = resolve_otel_config();
-                assert!(config.is_none(), "all-empty env vars should disable OTEL");
-            },
-        );
-    }
-
-    #[test]
-    fn trace_sampler_config_reads_and_clamps_sample_rate_overrides() {
-        with_test_env(
-            &[
-                ("TEMPER_TRACE_WASM_AUX_SAMPLE_PCT", Some("17")),
-                ("TEMPER_TRACE_DISPATCH_BACKGROUND_SAMPLE_PCT", Some("250")),
-            ],
-            || {
-                let config = TraceSamplerConfig::from_env();
-                assert_eq!(config.reduced_rule_rate("wasm_auxiliary"), Some(17));
-                assert_eq!(config.reduced_rule_rate("dispatch_background"), Some(100));
-                assert_eq!(
-                    config.reduced_prefix_count(),
-                    WASM_AUXILIARY_PREFIXES.len() + DISPATCH_BACKGROUND_PREFIXES.len()
-                );
-            },
-        );
-    }
-}
+mod tests;
