+++ scripts/verify-latency-observability-package.sh
@@ -69,8 +69,8 @@ require_pattern "$ROOT/crates/temper-store-postgres/src/metrics.rs" "temper_post
 require_pattern "$ROOT/crates/temper-server/src/query_projection_metrics.rs" "temper_query_projection_replay_parity_check_total" "projection replay parity metric"
 require_pattern "$ROOT/crates/temper-server/src/odata/read_support/shadow.rs" "TEMPER_ODATA_CATALOG_SHADOW_READ_EVERY" "projection shadow-read gate"
 require_regex "$ROOT/crates/temper-store-turso/src/schema/query_plane.rs" "fields[[:space:]]+TEXT NOT NULL DEFAULT '\\{\\}'" "Turso catalog fields column"
-require_pattern "$ROOT/crates/temper-observe/src/otel.rs" "temper_trace_sampler_decisions_total" "trace sampler decision metric"
-require_pattern "$ROOT/crates/temper-observe/src/otel.rs" "TEMPER_TRACE_DISPATCH_BACKGROUND_SAMPLE_PCT" "dispatch background trace budget flag"
+require_pattern "$ROOT/crates/temper-observe/src/otel/sampler.rs" "temper_trace_sampler_decisions_total" "trace sampler decision metric"
+require_pattern "$ROOT/crates/temper-observe/src/otel/sampler.rs" "TEMPER_TRACE_DISPATCH_BACKGROUND_SAMPLE_PCT" "dispatch background trace budget flag"
 require_pattern "$ROOT/crates/temper-server/src/trigger/dispatcher.rs" "reaction.fired_count" "reaction fanout summary span"
 require_pattern "$ROOT/crates/temper-authz/src/metrics.rs" "temper_cedar_evaluation_duration_ms" "Cedar duration ms metric"
 require_pattern "$ROOT/crates/temper-authz/src/metrics.rs" "temper_cedar_evaluation_phase_duration_ms" "Cedar phase duration metric"
